+++ typedefs.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class Status(Enum):
+    ongoing = 404
+    win_blue = 1
+    win_red = -1
+    draw = 0
+
