--- utils_setup.py
@@ -1,6 +1,0 @@
-from distutils.core import setup
-from Cython.Build import cythonize
-
-setup(
-    ext_modules=cythonize("utils.py")
-)
